Make the user session data pluggable via UserSessionDAO
Session data was reachable only through the relational UserSessionStore, so a
non-relational session store could not serve a login even after SessionDataStore
became pluggable.

Move the session-scoped operations of UserSessionStore into UserSessionDAO and
select the DAO by configuration, so that only session state is pluggable. The
user, identity provider and application data (IDN_AUTH_USER, IDP, SP_APP) stays
in UserSessionStore with a single implementation, so an alternative store neither
reimplements it nor duplicates the data.

UserSessionStore keeps every public method and its singleton accessor, and now
delegates the 27 session operations to the configured DAO. UserSessionDAOImpl
receives those JDBC bodies unchanged and remains the default; the new
UserSessionDAOFactory returns the configured DAO, and an alternative store
registers its own as an OSGi UserSessionDAO service named by
SessionStoreImplType. The new interface methods are default methods, so existing
implementations keep compiling.

Store selection is shared with SessionDataStore through SessionStorageSelector,
which owns the config property, the default store name and the case-insensitive
name matching. Selection fails rather than falling back to the relational store,
since a fallback would split session data between stores.
